--- a/.travis.yml
+++ b/.travis.yml
@@ -4,7 +4,8 @@ os:
   - linux
   - osx
 julia:
-  - release
+#  - release
+  - 0.6
   - nightly
 notifications:
   email: false