--- a/src/functions/distL2.jl
+++ b/src/functions/distL2.jl
@@ -2,7 +2,7 @@
 immutable DistL2{R <: Real, T <: ProximableFunction} <: ProximableFunction
   ind::T
   lambda::R
-  function DistL2(ind::T, lambda::R)
+  function DistL2{R, T}(ind::T, lambda::R) where {R <: Real, T <: ProximableFunction}
     if !is_set(ind) || !is_convex(ind)
       error("`ind` must be a convex set")
     end