--- a/src/functions/indBallL2.jl
+++ b/src/functions/indBallL2.jl
@@ -8,7 +8,7 @@ Returns the function `g = ind{x : ||x|| ⩽ r}`, for a real parameter `r > 0`.
 
 immutable IndBallL2{R <: Real} <: IndicatorConvex
   r::R
-  function IndBallL2(r::R)
+  function IndBallL2{R}(r::R) where {R <: Real}
     if r <= 0
       error("parameter r must be positive")
     else