--- a/src/functions/indFree.jl
+++ b/src/functions/indFree.jl
@@ -17,7 +17,7 @@ is_separable(f::IndFree) = true
 Returns the identically zero function.
 """
 
-typealias Zero IndFree
+const Zero = IndFree
 
 function (f::IndFree)(x::AbstractArray)
   return 0.0