--- a/src/functions/leastSquares.jl
+++ b/src/functions/leastSquares.jl
@@ -5,8 +5,6 @@
 
 Returns the function `f(x) = (λ/2)⋅||Ax-b||^2`.
 """
-<<<<<<< HEAD
-#TODO Fix abstract Array types, Union type, Real, and not Immutable?
 type LeastSquares{RC <: RealOrComplex, R<:Real, M<:AbstractArray{RC,2}, V<:AbstractArray{RC,1}, F<:Factorization} <: ProximableFunction
   A::M
   b::V
@@ -15,8 +13,7 @@ type LeastSquares{RC <: RealOrComplex, R<:Real, M<:AbstractArray{RC,2}, V<:Abstr
   gamma::R
   S::M
   U::F
-  function LeastSquares{RC,R,M,V,F}(A::M, b::V, lambda::R)
-      where {RC <: RealOrComplex, R<:Real, M<:AbstractArray{RC,2}, V<:AbstractArray{RC,1}, F<:Factorization}
+  function LeastSquares{RC,R,M,V,F}(A::M, b::V, lambda::R) where {RC <: RealOrComplex, R<:Real, M<:AbstractArray{RC,2}, V<:AbstractArray{RC,1}, F<:Factorization}
     if size(A, 1) != length(b)
       error("A and b have incompatible dimensions")
     end