--- a/src/functions/sqrDistL2.jl
+++ b/src/functions/sqrDistL2.jl
@@ -3,7 +3,7 @@
 immutable SqrDistL2{R <: Real} <: ProximableConvex
   ind::IndicatorConvex
   lambda::R
-  function SqrDistL2(ind::IndicatorConvex, lambda::R)
+  function SqrDistL2{R}(ind::IndicatorConvex, lambda::R) where {R <: Real}
     if lambda < 0
       error("parameter λ must be nonnegative")
     else