--- a/src/functions/sqrNormL2.jl
+++ b/src/functions/sqrNormL2.jl
@@ -2,7 +2,7 @@
 
 immutable SqrNormL2{T <: Union{Real, AbstractArray}} <: ProximableConvex
   lambda::T
-  function SqrNormL2(lambda::T)
+  function SqrNormL2{T}(lambda::T) where {T <: Union{Real,AbstractArray}}
     if any(lambda .< 0)
       error("coefficients in λ must be nonnegative")
     else