--- a/test/test_calls.jl
+++ b/test/test_calls.jl
@@ -30,6 +30,11 @@ stuff = [
         "args"   => ( randn(10), randn(20) )
       ),
 
+  Dict( "constr" => IndAffine,
+        "params" => ( (sprand(50,100, 0.1), randn(50)), (sprand(Complex{Float64}, 50,100, 0.1), randn(50)+im*randn(50)), ),
+        "args"   => ( randn(100), randn(100)+im*randn(100), )
+      ),
+
   Dict( "constr" => IndBallLinf,
         "wrong"  => ( (-rand(),), ),
         "params" => ( (rand(),), ),