Work in progress.
Fixes warnings on 0.6.
Fixes abstract types #23
More work to be done before pull.

**Notable performance examples:** taken from tests, observe that first calls include allocation

ProximalOperators.IndAffine
description : indicator of an affine subspace
domain      : AbstractArray{Real,1}, AbstractArray{Complex,1}
expression  : x ↦ 0 if Ax = b, +∞ otherwise
parameters  : A = Array{Float64,2} of size (1, 10), b = Array{Float64,1} of size (1,)

OLD:

* call                        :   **0.006850 seconds (2.19 k allocations: 106.342 KiB)**
* prox        :   **0.649409 seconds (515.29 k allocations: 24.416 MiB**, 0.88% gc time)
* prox!       :   **0.000016** seconds (15 allocations: 1.125 KiB)
* prox_naive  :   0.000011 seconds (17 allocations: 1.172 KiB)
* prox        :   **0.000005** seconds (17 allocations: 1.313 KiB)
* prox!       :   **0.000004** seconds (15 allocations: 1.125 KiB)
* prox_naive  :   0.000005 seconds (17 allocations: 1.172 KiB)

NEW:

* call                        :   **0.000004 seconds (2 allocations: 192 bytes)**
* prox        :   **0.000007 seconds (17 allocations: 1.313 KiB)**
* prox!       :   **0.000001** seconds (15 allocations: 1.125 KiB)
* prox_naive  :   0.000001 seconds (16 allocations: 1.156 KiB)
* prox        :   **0.000001** seconds (17 allocations: 1.313 KiB)
* prox!       :   **0.000001** seconds (15 allocations: 1.125 KiB)
* prox_naive  :   0.000001 seconds (16 allocations: 1.156 KiB)

ProximalOperators.LeastSquares
description : least-squares penalty
domain      : AbstractArray{Real,1}
expression  : x ↦ (λ/2)||A*x - b||^2
parameters  : λ = 1.0, A = Array{Float64,2} of size (10, 25), b = Array{Float64,1} of size (10,)
OLD:

* call                        :   **0.002280 seconds (17 allocations: 1.266 KiB)**
* prox        :   **0.039139 seconds (14.41 k allocations: 705.142 KiB)**
* prox!       :   **0.000055** seconds (17 allocations: 2.391 KiB)
* prox_naive  :   0.004583 seconds (47 allocations: 19.031 KiB)
* prox        :   **0.000046 seconds (32 allocations: 4.734 KiB)**
* prox!       :   0.000007 seconds (17 allocations: 2.391 KiB)
* prox_naive  :   0.000049 seconds (22 allocations: 17.063 KiB)

NEW:
* call                        :   **0.000005 seconds (2 allocations: 320 bytes)**
* prox        :   **0.000054 seconds (20 allocations: 4.547 KiB)**
* prox!       :   **0.000004** seconds (10 allocations: 2.281 KiB)
* prox_naive  :   0.006062 seconds (14 allocations: 16.938 KiB)
* prox        :   **0.000009 seconds (20 allocations: 4.547 KiB)**
* prox!       :   0.000004 seconds (10 allocations: 2.281 KiB)
* prox_naive  :   0.000029 seconds (14 allocations: 16.938 KiB)

ProximalOperators.LeastSquares
description : least-squares penalty
domain      : AbstractArray{Real,1}
expression  : x ↦ (λ/2)||A*x - b||^2
parameters  : λ = 0.2654585810275454, A = Array{Float64,2} of size (40, 13), b = Array{Float64,1} of size (40,)
OLD:
* call                        :   0.000006 seconds (6 allocations: 960 bytes)
* prox        :   **0.000031** seconds (28 allocations: 4.922 KiB)
* prox!       :   0.000005 seconds (13 allocations: 1.563 KiB)
* prox_naive  :   0.000019 seconds (22 allocations: 6.234 KiB)
* prox        :   **0.000016** seconds (28 allocations: 4.922 KiB)
* prox!       :   0.000004 seconds (13 allocations: 1.563 KiB)
* prox_naive  :   0.000015 seconds (22 allocations: 6.234 KiB)

NEW:
* call                        :   0.000002 seconds (2 allocations: 896 bytes)
* prox        :   **0.000010** seconds (16 allocations: 4.734 KiB)
* prox!       :   0.000003 seconds (6 allocations: 1.453 KiB)
* prox_naive  :   0.000014 seconds (14 allocations: 6.109 KiB)
* prox        :   **0.000007** seconds (16 allocations: 4.734 KiB)
* prox!       :   0.000003 seconds (6 allocations: 1.453 KiB)
* prox_naive  :   0.000011 seconds (14 allocations: 6.109 KiB)

ProximalOperators.LeastSquares
description : least-squares penalty
domain      : AbstractArray{Complex,1}
expression  : x ↦ (λ/2)||A*x - b||^2
parameters  : λ = 1.0, A = Array{Complex{Float64},2} of size (25, 10), b = Array{Complex{Float64},1} of size (25,)

OLD:
* call                        :   **0.004213 seconds (2.00 k allocations: 94.550 KiB)**
* prox        :   **0.054817 seconds (15.56 k allocations: 787.318 KiB)**
* prox!       :   **0.000040** seconds (13 allocations: 1.891 KiB)
* prox_naive  :   **0.666595 seconds (450.84 k allocations: 20.933 MiB)**
* prox        :   **0.000052** seconds (28 allocations: 5.922 KiB)
* prox!       :   **0.000009** seconds (13 allocations: 1.891 KiB)
* prox_naive  :   **0.000037** seconds (22 allocations: 7.469 KiB)

NEW:
* call                        :   **0.000025 seconds (2 allocations: 1.063 KiB)**
* prox        :   **0.000071 seconds (16 allocations: 5.734 KiB)**
* prox!       :   **0.000005** seconds (6 allocations: 1.781 KiB)
* prox_naive  :   **0.000034 seconds (14 allocations: 7.344 KiB)**
* prox        :   **0.000022** seconds (16 allocations: 5.734 KiB)
* prox!       :   **0.000003** seconds (6 allocations: 1.781 KiB)
* prox_naive  :   **0.000020** seconds (14 allocations: 7.344 KiB)

----------------------------------------------------------
ProximalOperators.LeastSquares
description : least-squares penalty
domain      : AbstractArray{Real,1}
expression  : x ↦ (λ/2)||A*x - b||^2
parameters  : λ = 0.3611825580251262, A = SparseMatrixCSC{Float64,Int64} of size (100, 1000), b = Array{Float64,1} of size (100,)

OLD:
* call                        :   **0.037073 seconds (11.04 k allocations: 566.955 KiB)**
* prox        :   **0.330995** seconds (100.69 k allocations: 5.369 MiB)
* prox!       :   0.000107 seconds (27 allocations: 46.039 KiB)
* prox_naive  :   **0.732980** seconds (390.98 k allocations: 51.149 MiB, 2.35% gc time)
* prox        :   0.000703 seconds (95 allocations: 830.305 KiB)
* prox!       :   0.000107 seconds (27 allocations: 46.039 KiB)
* prox_naive  :   0.109704 seconds (5.14 k allocations: 33.245 MiB, 17.00% gc time)

NEW:
* call                        :   **0.000046 seconds (2 allocations: 1.750 KiB)**
* prox        :   **0.134826** seconds (24.31 k allocations: 1.987 MiB)
* prox!       :   0.000098 seconds (19 allocations: 45.914 KiB)
* prox_naive  :   **0.243142** seconds (38.70 k allocations: 34.512 MiB, 2.21% gc time)
* prox        :   0.000713 seconds (83 allocations: 826.703 KiB)
* prox!       :   0.000104 seconds (19 allocations: 45.914 KiB)
* prox_naive  :   0.109826 seconds (5.15 k allocations: 33.026 MiB, 20.42% gc time)