@lostella 2017-04-08
Check out also #27, you may want to merge it into your branch since it touches almost all functions.