@lostella 2017-04-08
The speedups look impressive! We should probably set up a bunch of fixed set of cases to use as benchmarks, at least for the most important functions, and measure the time required to `prox!' using `@benchmark' and other utilities from BenchmarkTools.jl.