@mfalt 2017-04-08
Yes, I agree, they have something similar for the julia repo, we might want
to take a look at that.

On Sat, Apr 8, 2017, 22:17 Lorenzo Stella <notifications@github.com> wrote:

> The speedups look impressive! We should probably set up a bunch of fixed
> set of cases to use as benchmarks, at least for the most important
> functions, and measure the time required to prox!' using@benchmark
> <https://github.com/benchmark>' and other utilities from
> BenchmarkTools.jl.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/kul-forbes/ProximalOperators.jl/pull/26#issuecomment-292742935>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ADFlSVIqI16heEbew-KXkRKDd0dPWx1fks5rt-s_gaJpZM4M3gfr>
> .
>