@lostella 2017-05-26
I have opened #35 to propose some changes to `SeparableSum` and `SlicedSeparableSum`. If we all agree on those changes we can probably merge #35, then resolve all conflicts here and probably merge also #26?