@lostella 2017-05-26
The other option is that I push those modifications directly on this branch as new commits.