@nantonel 2017-05-26
I'm in favour of the last option:

> EDIT: or maybe, we first merge #26 and then verify how #35 fits into that. Maybe this is better.